+++ src/masonite/authentication/Auth.py
@@ -1,5 +1,6 @@
 import pendulum
 import uuid
+from ..routes import Route
 
 
 class Auth:
@@ -132,3 +133,22 @@ class Auth:
             .where("token", token)
             .delete()
         )
+
+    @classmethod
+    def routes(self):
+        Route.get("/login", "auth.LoginController@show").name("login")
+        Route.get("/home", "auth.HomeController@show").name("auth.home")
+        Route.get("/register", "auth.RegisterController@show").name("register")
+        Route.post("/register", "auth.RegisterController@store").name("register.store")
+        Route.get("/password_reset", "auth.PasswordResetController@show").name("password_reset")
+        Route.post("/password_reset", "auth.PasswordResetController@store").name(
+            "password_reset.store"
+        )
+        Route.get("/change_password", "auth.PasswordResetController@change_password").name(
+            "change_password"
+        )
+        Route.post(
+            "/change_password", "auth.PasswordResetController@store_changed_password"
+        ).name("change_password.store")
+        Route.post("/login", "auth.LoginController@store").name("login.store")
+        Route.post("/login", "auth.LoginController@store").name("login.store")
