+++ src/masonite/broadcasting/BroadcastRoutes.py
@@ -11,8 +11,4 @@ class BroadcastRoutes:
         )
 
 
-class BroadcastingController:
-    def authorize(self, request: Request, broadcast: Broadcast):
-        return broadcast.driver("pusher").authorize(
-            request.input("channel_name"), request.input("socket_id")
-        )
+
