+++ src/masonite/broadcasting/controllers/BroadcastingController.py
@@ -1,0 +1,5 @@
+class BroadcastingController:
+    def authorize(self, request: Request, broadcast: Broadcast):
+        return broadcast.driver("pusher").authorize(
+            request.input("channel_name"), request.input("socket_id")
+        )
