+++ src/masonite/broadcasting/controllers/__init__.py
@@ -1,0 +1,1 @@
+from .BroadcastingController import BroadcastingController
